nutrition_tracker/requirements.txt
Flask==2.3.2
pymongo==4.5.0
PyJWT==2.8.0
python-dotenv==1.0.0
werkzeug==3.0.0
